requirements.txt
flet==0.22.*
python-dotenv==1.0.1
mysql-connector-python == 8.3.*